everything now works with an atomic start time property. but now I will get rid of this and use current frame instead

+++ libraries/entities-renderer/src/RenderableModelEntityItem.h
@@ -187,7 +187,7 @@ private:
     float _currentFrame { 0 };
     float _endAnim{ 0 };
     bool tempbool{ false };
-    float _currentlyPlayingFrame{ 0 };
+    quint64 _currentlyPlayingFrame{ 0 };
 };
 
 } } // namespace
